fix: ensure search paths will be indexed (on Windows)
The file search extension relies on the OS's desktop search to work, and it
is possible that the desktop search indexer may not index the search paths
we specify.

This commit fixes the issue by adding a hook for configuration
changes. Within this hook, we instruct the indexer to index the search paths
defined in the configuration.

Currently, this is implemented for Windows. Similar implementations will be
added for other platforms in future commits.